+++ core/src/main/java/com/lgdtimtou/customenchantments/enchantments/created/fields/CustomEnchantLevel.java
@@ -2,6 +2,6 @@ package com.lgdtimtou.customenchantments.enchantments.created.fields;
 
 import java.util.Queue;
 
-public record CustomEnchantLevel(int cooldown, String cooldownMessage, double chance, boolean cancelEvent,
+public record CustomEnchantLevel(double cooldown, String cooldownMessage, double chance, boolean cancelEvent,
                                  Queue<CustomEnchantInstruction> instructions) {
 }
